Refactor Modal Component
Styled Components and other assets got ported to their own file so Modal
is leaner and ready for additional Modes.
